+++ server/package.json
@@ -28,6 +28,7 @@
     "body-parser": "^1.14.1",
     "dotenv": "^1.2.0",
     "express": "^4.13.3",
+    "jsonschema": "^1.0.2",
     "mysql": "^2.9.0",
     "sequelize": "^3.14.2"
   }
